update doc with checkov support
