+++ src/main/java/com/hack23/sonar/cloudformation/parser/cfnnag/CfnNagReport.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package com.hack23.sonar.cloudformation.parser;
+package com.hack23.sonar.cloudformation.parser.cfnnag;
 
 import java.util.ArrayList;
 import java.util.List;
