+++ src/test/java/com/hack23/sonar/cloudformation/parser/checkov/CheckovSonarqubeRuleGeneratorTest.java
@@ -25,7 +25,6 @@ import java.nio.charset.StandardCharsets;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
 import org.apache.commons.csv.CSVFormat;
 import org.apache.commons.csv.CSVParser;
